requirements.txt
cassandra-driver
fastapi
uvicorn
python-dotenv
requests
python-multipart